+++ src/Data/Conduit/Tar/Unix.hs
@@ -1,3 +1,4 @@
+{-# LANGUAGE CPP                 #-}
 {-# LANGUAGE OverloadedStrings   #-}
 {-# LANGUAGE RecordWildCards     #-}
 {-# LANGUAGE ScopedTypeVariables #-}
@@ -7,18 +8,21 @@ module Data.Conduit.Tar.Unix
     ) where
 
 import           Conduit
-import           Control.Exception
-import           Control.Monad                 (when)
+import           Control.Exception.Safe
+import           Control.Monad                 (void, when)
 import           Data.Bits
 import qualified Data.ByteString.Char8         as S8
-import           Data.Conduit.Tar.Types        (FileInfo (..), FileType (..))
+import           Data.Conduit.Tar.Types
+import           Foreign.C.Types               (CTime (..))
 import qualified System.Directory              as Dir
-import qualified System.Posix.Files.ByteString as Posix
+import qualified System.Posix.Files            as Posix
 import qualified System.Posix.User             as Posix
 
-getFileInfo :: S8.ByteString -> IO FileInfo
-getFileInfo fp = do
-    fs <- Posix.getSymbolicLinkStatus fp
+
+getFileInfo :: FilePath -> IO FileInfo
+getFileInfo fpStr = do
+    let fp = encodeFilePath fpStr
+    fs <- Posix.getSymbolicLinkStatus fpStr
     let uid = Posix.fileOwner fs
         gid = Posix.fileGroup fs
     -- Allow for username/group retrieval failure, especially useful for non-tty environment.
@@ -30,14 +34,14 @@ getFileInfo fp = do
         case () of
             () | Posix.isRegularFile fs     -> return (FTNormal, Posix.fileSize fs)
                | Posix.isSymbolicLink fs    -> do
-                     ln <- Posix.readSymbolicLink fp
-                     return (FTSymbolicLink ln, 0)
+                     ln <- Posix.readSymbolicLink fpStr
+                     return (FTSymbolicLink (encodeFilePath ln), 0)
                | Posix.isCharacterDevice fs -> return (FTCharacterSpecial, 0)
                | Posix.isBlockDevice fs     -> return (FTBlockSpecial, 0)
                | Posix.isDirectory fs       -> return (FTDirectory, 0)
                | Posix.isNamedPipe fs       -> return (FTFifo, 0)
                | otherwise                  -> error $ "Unsupported file type: " ++ S8.unpack fp
-    return FileInfo
+    return $! FileInfo
         { filePath      = fp
         , fileUserId    = uid
         , fileUserName  = either (const "") (S8.pack . Posix.userName) euEntry
@@ -49,27 +53,41 @@ getFileInfo fp = do
         , fileModTime   = Posix.modificationTime fs
         }
 
+
 -- | Restore files onto the file system. Produces actions that will set the modification time on the
 -- directories, which can be executed after the pipeline has finished and all files have been
 -- written to disk.
 restoreFile :: (MonadResource m) =>
                FileInfo -> ConduitM S8.ByteString (IO ()) m ()
 restoreFile FileInfo {..} = do
-    let filePath' = S8.unpack filePath
+    let fpStr = decodeFilePath filePath
+        restorePermissions = do
+            void $ tryAny $ Posix.setOwnerAndGroup fpStr fileUserId fileGroupId
+            void $ tryAny $ Posix.setFileMode fpStr fileMode
     case fileType of
         FTDirectory -> do
-            liftIO $ Dir.createDirectoryIfMissing False filePath'
+            liftIO $ do
+                Dir.createDirectoryIfMissing False fpStr
+                restorePermissions
             yield $
-                (Dir.doesDirectoryExist filePath' >>=
-                 (`when` Posix.setFileTimes filePath fileModTime fileModTime))
+                (Dir.doesDirectoryExist fpStr >>=
+                 (`when` Posix.setFileTimes fpStr fileModTime fileModTime))
         FTSymbolicLink link ->
             liftIO $ do
-                exist <- Posix.fileExist filePath
-                when exist $ Dir.removeFile filePath'
-                Posix.createSymbolicLink link filePath
-        FTNormal -> sinkFile filePath'
+                -- Try to unlink any existing file/symlink
+                _ <- tryAny $ Posix.removeLink fpStr
+                Posix.createSymbolicLink (decodeFilePath link) fpStr
+                _ <- tryAny $ Posix.setSymbolicLinkOwnerAndGroup fpStr fileUserId fileGroupId
+                -- Try best effort in setting symbolic link modification time.
+#if MIN_VERSION_unix(2,7,0)
+                let CTime epochInt32 = fileModTime
+                    unixModTime = fromInteger (fromIntegral epochInt32)
+                void $ tryAny $ Posix.setSymbolicLinkTimesHiRes fpStr unixModTime unixModTime
+#endif
+        FTNormal -> do
+            sinkFile fpStr
+            liftIO $ do
+                restorePermissions
+                Posix.setFileTimes fpStr fileModTime fileModTime
         ty -> error $ "Unsupported tar entry type: " ++ show ty
-    liftIO $ do
-        Posix.setFileTimes filePath fileModTime fileModTime
-        Posix.setSymbolicLinkOwnerAndGroup filePath fileUserId fileGroupId
-        Posix.setFileMode filePath fileMode
+